Split source map composer helpers into focused modules
Move line/column conversion into `line-index.ts` and atom-range
translation into `atom-translator.ts`. Each helper was previously
exported solely so its tests could reach it, which `knip --production`
flagged as unused exports. With the helpers in their own modules and
consumed by `source-map-composer.ts` itself, the exports have a real
production consumer and the test seams stay narrow.
